add version check for scipy when simplex is used

+++ examples/IMPACT-Z/impact-z_single_simplex.py
@@ -112,7 +112,17 @@ def objectives(point, *args):
 
 	return retval
 
-			
+def compareVersion(version1, version2):
+	versions1 = [int(v) for v in version1.split(".")]
+	versions2 = [int(v) for v in version2.split(".")]
+	for i in range(max(len(versions1),len(versions2))):
+		v1 = versions1[i] if i < len(versions1) else 0
+		v2 = versions2[i] if i < len(versions2) else 0
+		if v1 > v2:
+			return 1
+		elif v1 <v2:
+			return -1
+	return 0
 def main():
 
 	# Parse command line arguments
@@ -176,6 +186,8 @@ def main():
 
 	Pdefault = np.asarray([0,0,0,0,0])
 	if(TUNER_NAME=='SIMPLEX'):
+		if(compareVersion(sp.__version__,'1.7.0')==-1):
+			raise Exception(f"Simplex require scipy version >=1.7.0 to support bounds")
 		os.system("rm -rf history.csv")
 
 		lw = [-1]*Pdefault.shape[0]
